+++ R/euclideanPath.R
@@ -448,12 +448,23 @@ print.euclidean_path <- function(x, ...) {
 }
 
 casePumpEucl <- function(orgn, orgn.nm, destination, dstn, latlong, pmp,
-  vestry, case.set, location) {
+  vestry, case.set, location, square.intersections) {
 
   if (latlong) vars <- c("lon", "lat")
   else vars <- c("x", "y")
 
   if (case.set == "observed") {
+    vars.lndmrk <- c("case", vars, "name")
+    lndmrk.sq <- cholera::landmark.squaresB[, vars.lndmrk]
+
+    if (square.intersections) {
+      lndmrk <- rbind(lndmrk.sq, cholera::landmarksB[, vars.lndmrk])
+    } else {
+      sq.intersections <- grep("Square", cholera::landmarksB$name)
+      lndmrk.etc <- cholera::landmarksB[-sq.intersections, vars.lndmrk]
+      lndmrk <- rbind(lndmrk.sq, lndmrk.etc[, vars.lndmrk])
+    }
+
     if (location %in% c("nominal", "anchor")) {
       if (any(orgn < 1000L)) {
         if (location == "anchor") {
@@ -465,15 +476,14 @@ casePumpEucl <- function(orgn, orgn.nm, destination, dstn, latlong, pmp,
       }
 
       fatal <- cholera::fatalities$case %in% orgn
-      land <- cholera::landmarksB$case %in% orgn
-      # sq <- cholera::landmark.squaresB$case %in% orgn  # via validateCase()!
+      land <- lndmrk$case %in% orgn
 
       if (any(fatal) & any(land)) {
         a <- cholera::fatalities[fatal, vars]
-        b <- cholera::landmarksB[land, vars]
+        b <- lndmrk[land, vars]
         ego.coords <- rbind(a, b)
       } else if (all(!fatal) & any(land)) {
-        ego.coords <- cholera::landmarksB[land, vars]
+        ego.coords <- lndmrk[land, vars]
       } else if (any(fatal) & all(!land)) {
         ego.coords <- cholera::fatalities[fatal, vars]
       }
@@ -485,20 +495,20 @@ casePumpEucl <- function(orgn, orgn.nm, destination, dstn, latlong, pmp,
         orgn <- cholera::anchor.case[sel, "anchor"]
 
         fatal <- ortho$case %in% orgn
-        land <- cholera::landmarksB$case %in% orgn
+        land <- lndmrk$case %in% orgn
       } else {
         ortho <- cholera::ortho.proj
         names(ortho)[names(ortho) %in% paste0(vars, ".proj")] <- vars
         fatal <- ortho$case %in% orgn
-        land <- cholera::landmarksB$case %in% orgn
+        land <- lndmrk$case %in% orgn
       }
 
       if (any(fatal) & any(land)) {
         a <- ortho[fatal, vars]
-        b <- cholera::landmarksB[land, vars]
+        b <- lndmrk[land, vars]
         ego.coords <- rbind(a, b)
       } else if (all(!fatal) & any(land)) {
-        ego.coords <- cholera::landmarksB[land, vars]
+        ego.coords <- lndmrk[land, vars]
       } else if (any(fatal) & all(!land)) {
         ego.coords <- ortho[fatal, vars]
       }
@@ -524,34 +534,60 @@ casePumpEucl <- function(orgn, orgn.nm, destination, dstn, latlong, pmp,
 
   alter.coords <- pmp[pmp$id %in% dstn, vars]
 
-  sel <- seq_len(nrow(alter.coords))
+  if (nrow(ego.coords) == 1 & nrow(alter.coords) == 1) {
+    d <- stats::dist(rbind(ego.coords, alter.coords))
+    alter.id <- 1L
+    nearest.pump <- dstn
+    ego <- ego.coords
+    alter <- alter.coords
 
-  if (nrow(ego.coords) == 1) {
-    d <- stats::dist(rbind(ego.coords, alter.coords))[sel]
-    d.sel <- which.min(d)
+  } else if (nrow(ego.coords) == 1 & nrow(alter.coords) > 1) {
+    d.sel <- seq_len(nrow(alter.coords))
+    ds <- stats::dist(rbind(ego.coords, alter.coords))[d.sel]
 
-    nearest.pump <- pmp[d.sel, "id"]
-    nearest.d <- min(d)
+    d <- min(ds)
+    alter.id <- which.min(ds)
+    nearest.pump <- pmp[alter.id, "id"]
 
     ego <- ego.coords
-    alter <- alter.coords[d.sel, ]
-  } else if (length(ego.coords) > 1) {
+    alter <- alter.coords[alter.id, ]
+
+  } else if (nrow(ego.coords) > 1 & nrow(alter.coords) == 1) {
+    ds <- vapply(seq_len(nrow(ego.coords)), function(i) {
+      stats::dist(rbind(ego.coords[i, ], alter.coords))
+    }, numeric(1L))
+
+    d <- min(ds)
+    alter.id <- 1L
+    nearest.pump <- dstn
+
+    ego.id <- which.min(ds)
+    orgn <- orgn[ego.id]
+    orgn.nm <- orgn.nm[ego.id]
+
+    ego <- ego.coords[ego.id, ]
+    alter <- alter.coords
+
+
+
+  } else if (nrow(ego.coords) > 1 & nrow(alter.coords) > 1) {
     d.multi.ego <- lapply(seq_len(nrow(ego.coords)), function(i) {
       stats::dist(rbind(ego.coords[i, ], alter.coords))[sel]
     })
+
     ego.id <- which.min(vapply(d.multi.ego, min, numeric(1L)))
     orgn <- orgn[ego.id]
     orgn.nm <- orgn.nm[ego.id]
+
     d <- d.multi.ego[[ego.id]]
-    nearest.pump <- dstn[which.min(d)]
-    nearest.d <- min(d)
+    alter.id <- which.min(d)
 
     ego <- ego.coords[ego.id, ]
-    alter <- alter.coords[which.min(d), ]
+    alter <- alter.coords[alter.id, ]
   }
 
   data.summary <- data.frame(orgn = orgn, orgn.nm = orgn.nm,
-    nearest.dest = nearest.pump, d = nearest.d)
+    nearest.dest = dstn[alter.id], d = d)
 
   list(ego = ego, alter = alter, data = data.summary)
 }